+++ fdbclient/CMakeLists.txt
@@ -65,7 +65,6 @@ set(FDBCLIENT_SRCS
   GlobalConfig.actor.cpp
   GrvProxyInterface.h
   HighContentionPrefixAllocator.actor.h
-  HTTP.actor.cpp
   IClientApi.h
   IConfigTransaction.cpp
   IConfigTransaction.h
@@ -106,8 +105,6 @@ set(FDBCLIENT_SRCS
   SpecialKeySpace.actor.h
   ReadYourWrites.actor.cpp
   ReadYourWrites.h
-  RESTClient.h
-  RESTClient.actor.cpp
   RestoreInterface.cpp
   RestoreInterface.h
   RunTransaction.actor.h
