+++ flow/Knobs.h
@@ -343,6 +343,14 @@ public:
 	int64_t ENCRYPT_CIPHER_KEY_CACHE_TTL;
 	int64_t ENCRYPT_KEY_REFRESH_INTERVAL;
 
+	// RESTClient
+	int RESTCLIENT_MAX_CONNECTIONPOOL_SIZE;
+	int RESTCLIENT_CONNECT_TRIES;
+	int RESTCLIENT_CONNECT_TIMEOUT;
+	int RESTCLIENT_MAX_CONNECTION_LIFE;
+	int RESTCLIENT_REQUEST_TRIES;
+	int RESTCLIENT_REQUEST_TIMEOUT_SEC;
+
 	FlowKnobs(class Randomize, class IsSimulated);
 	void initialize(class Randomize, class IsSimulated);
 };
