+++ flow/error_definitions.h
@@ -148,6 +148,7 @@ ERROR( http_auth_failed, 1524, "HTTP request failed due to bad credentials" )
 ERROR( http_bad_request_id, 1525, "HTTP response contained an unexpected X-Request-ID header" )
 ERROR( rest_invalid_uri, 1526, "Invalid REST URI")
 ERROR( rest_invalid_rest_client_knob, 1527, "Invalid RESTClient knob")
+ERROR( rest_connectpool_key_not_found, 1528, "ConnectKey not found in connection pool")
 
 
 // 2xxx Attempt (presumably by a _client_) to do something illegal.  If an error is known to
@@ -310,10 +311,6 @@ ERROR( encrypt_key_ttl_expired, 2703, "Expected encryption key TTL has expired")
 ERROR( encrypt_header_authtoken_mismatch, 2704, "Encryption header authentication token mismatch")
 ERROR( encrypt_update_cipher, 2705, "Attempt to update encryption cipher key")
 ERROR( encrypt_invalid_id, 2706, "Invalid encryption cipher details")
-
-// 28XX - RESTClient errors
-ERROR( invalid_rest_uri, 2800, "Invalid REST URI")
-ERROR( invalid_rest_client_knob, 2801, "Invalid RESTClient knob")
 
 // 4xxx Internal errors (those that should be generated only by bugs) are decimal 4xxx
 ERROR( unknown_error, 4000, "An unknown error occurred" )  // C++ exception not of type Error
